Added Complete Feature Integration Tests and added preclusion discalimer to user guide
